Refactor classes to have better separation
