+++ app/src/main/java/com/boswelja/contactringtonegenerator/contactpicker/StructuredName.kt
@@ -1,0 +1,9 @@
+package com.boswelja.contactringtonegenerator.contactpicker
+
+data class StructuredName(
+    val prefix: String?,
+    val firstName: String?,
+    val middleName: String?,
+    val lastName: String?,
+    val suffix: String?
+)
