--- app/src/main/java/com/boswelja/contactringtonegenerator/contacts/StructuredName.kt
@@ -1,9 +1,0 @@
-package com.boswelja.contactringtonegenerator.contacts
-
-data class StructuredName(
-    val prefix: String?,
-    val firstName: String?,
-    val middleName: String?,
-    val lastName: String?,
-    val suffix: String?
-)
