+++ app/src/main/java/com/boswelja/contactringtonegenerator/result/Result.kt
@@ -1,0 +1,8 @@
+package com.boswelja.contactringtonegenerator.result
+
+enum class Result {
+    FAILED,
+    MIXED,
+    SUCCESSFUL,
+    UNKNOWN
+}
