+++ app/src/main/java/com/boswelja/contactringtonegenerator/result/ui/ResultViewModelFactory.kt
@@ -1,0 +1,15 @@
+package com.boswelja.contactringtonegenerator.result.ui
+
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProvider
+
+class ResultViewModelFactory(
+    private val successCount: Int,
+    private val failCount: Int
+) : ViewModelProvider.Factory {
+
+    override fun <T : ViewModel?> create(modelClass: Class<T>): T {
+        return modelClass.getConstructor(Int::class.java, Int::class.java)
+            .newInstance(successCount, failCount)
+    }
+}
