+++ app/src/main/java/com/boswelja/contactringtonegenerator/ringtonegen/tts/SynthesisJob.kt
@@ -1,0 +1,17 @@
+package com.boswelja.contactringtonegenerator.ringtonegen.tts
+
+class SynthesisJob(val id: String, val text: String) {
+
+    override fun equals(other: Any?): Boolean {
+        if (other is SynthesisJob) {
+            return other.id == id
+        }
+        return super.equals(other)
+    }
+
+    override fun hashCode(): Int {
+        var result = id.hashCode()
+        result = 31 * result + text.hashCode()
+        return result
+    }
+}
