+++ app/src/main/java/com/boswelja/contactringtonegenerator/ringtonegen/tts/SynthesisResult.kt
@@ -1,0 +1,19 @@
+package com.boswelja.contactringtonegenerator.ringtonegen.tts
+
+import java.io.File
+
+data class SynthesisResult(val id: String, val result: File) {
+
+    override fun equals(other: Any?): Boolean {
+        if (other is SynthesisResult) {
+            return other.id == id
+        }
+        return super.equals(other)
+    }
+
+    override fun hashCode(): Int {
+        var result = id.hashCode()
+        result = 31 * result + this.result.hashCode()
+        return result
+    }
+}
