--- app/src/main/java/com/boswelja/contactringtonegenerator/tts/SynthesisResult.kt
@@ -1,19 +1,0 @@
-package com.boswelja.contactringtonegenerator.tts
-
-import java.io.File
-
-data class SynthesisResult(val id: String, val result: File) {
-
-    override fun equals(other: Any?): Boolean {
-        if (other is SynthesisResult) {
-            return other.id == id
-        }
-        return super.equals(other)
-    }
-
-    override fun hashCode(): Int {
-        var result = id.hashCode()
-        result = 31 * result + this.result.hashCode()
-        return result
-    }
-}
