--- app/src/main/java/com/boswelja/contactringtonegenerator/ui/results/Result.kt
@@ -1,8 +1,0 @@
-package com.boswelja.contactringtonegenerator.ui.results
-
-enum class Result {
-    FAILED,
-    MIXED,
-    SUCCESSFUL,
-    UNKNOWN
-}
