--- app/src/main/java/com/boswelja/contactringtonegenerator/ui/ringtonecreator/PickRingtone.kt
@@ -1,24 +1,0 @@
-package com.boswelja.contactringtonegenerator.ui.ringtonecreator
-
-import android.app.Activity
-import android.content.Context
-import android.content.Intent
-import android.media.RingtoneManager
-import android.net.Uri
-import androidx.activity.result.contract.ActivityResultContract
-
-class PickRingtone : ActivityResultContract<Int, Uri?>() {
-    override fun createIntent(context: Context, ringtoneType: Int) =
-        Intent(RingtoneManager.ACTION_RINGTONE_PICKER).apply {
-            putExtra(RingtoneManager.EXTRA_RINGTONE_SHOW_DEFAULT, true)
-            putExtra(RingtoneManager.EXTRA_RINGTONE_SHOW_SILENT, false)
-            putExtra(RingtoneManager.EXTRA_RINGTONE_TYPE, ringtoneType)
-        }
-
-    override fun parseResult(resultCode: Int, result: Intent?): Uri? {
-        if (resultCode != Activity.RESULT_OK) {
-            return null
-        }
-        return result?.getParcelableExtra(RingtoneManager.EXTRA_RINGTONE_PICKED_URI)
-    }
-}
