+++ app/src/test/java/com/boswelja/contactringtonegenerator/contacts/ContactTest.kt
@@ -1,5 +1,6 @@
 package com.boswelja.contactringtonegenerator.contacts
 
+import com.boswelja.contactringtonegenerator.contactpicker.Contact
 import org.junit.Assert.* // ktlint-disable
 import org.junit.Test
